Add new base view, use on homepage.
This adds a new base view, based on the US Web Design System, which is then used on the homepage to render the new acquisition form and a list of your acquisitions.

Additionally, switched the new acquisition form to accept a GET parameter instead of a URL parameter, which makes it easier to work with from the web (no JavaScript is required).